Move examples into separate file
